+++ src/service/artist/Artist.ts
@@ -52,16 +52,16 @@ export default class Artist {
       userId,
       artist,
     };
-    this.httpClient.post("bookmark", {
-      body
+    this.httpClient.post(`bookmark/${userId}/${artist._id}`, {
+      body,
     });
   }
 
   async getBookmark(userId: string, artistId: string) {
-    return this.httpClient.get("bookmark", {
+    return this.httpClient.get(`bookmark/${userId}/${artistId}`, {
       params: {
         userId,
-        artistId
+        artistId,
       },
     });
   }
@@ -71,7 +71,7 @@ export default class Artist {
   }
 
   async deleteBookmark(userId: string, artistId: string) {
-    this.httpClient.delete("bookmark", {
+    this.httpClient.delete(`bookmark/${userId}/${artistId}`, {
       params: {
         userId,
         artistId,
